Continue working on a primitive node
Scraping the scaffold from a primitive node would be a massive boon, but I need to rethink my
approach. There is definitely some overlap with the parser logic in that string literals will need
to be able to search for the string value in what's left. Otherwise, we would strip everything out
of a string like `  "    ",\n`, which is wrong.